chore: Update vendored sources (tag v1.0.0) to duckdb/duckdb@1f98600c2cf8722a6d2f2d805bb4af5e701319fc

+++ src/duckdb/src/include/duckdb/common/types/datetime.hpp
@@ -24,7 +24,7 @@ struct dtime_t { // NOLINT
 		return micros;
 	}
 	explicit inline operator double() const {
-		return static_cast<double>(micros);
+		return micros;
 	}
 
 	// comparison operators
